fix: 移除管理员权限逻辑修复

+++ src/main/java/com/wolfhouse/wolfhouseblog/service/impl/AuthorityServiceImpl.java
@@ -10,7 +10,6 @@ import com.wolfhouse.wolfhouseblog.common.utils.ServiceUtil;
 import com.wolfhouse.wolfhouseblog.mapper.AdminAuthMapper;
 import com.wolfhouse.wolfhouseblog.mapper.AuthorityMapper;
 import com.wolfhouse.wolfhouseblog.pojo.domain.Authority;
-import com.wolfhouse.wolfhouseblog.pojo.domain.table.AdminAuthTableDef;
 import com.wolfhouse.wolfhouseblog.pojo.domain.table.AuthorityTableDef;
 import com.wolfhouse.wolfhouseblog.pojo.dto.AuthorityByIdDto;
 import com.wolfhouse.wolfhouseblog.service.AuthorityService;
@@ -21,6 +20,8 @@ import org.springframework.util.CollectionUtils;
 
 import java.util.List;
 import java.util.Set;
+
+import static com.wolfhouse.wolfhouseblog.pojo.domain.table.AdminAuthTableDef.ADMIN_AUTH;
 
 /**
  * @author linexsong
@@ -64,9 +65,14 @@ public class AuthorityServiceImpl extends ServiceImpl<AuthorityMapper, Authority
     @Override
     public Boolean deleteAuthorityByIds(AuthorityByIdDto dto) {
         Set<Long> ids = dto.getIds();
+        Long admin = dto.getAdminId();
+        if (BeanUtil.isAnyBlank(ids, admin)) {
+            throw new ServiceException(AuthExceptionConstant.BAD_REQUEST);
+        }
         if (adminAuthMapper.deleteByQuery(
              QueryWrapper.create()
-                         .where(AdminAuthTableDef.ADMIN_AUTH.AUTH_ID.in(ids))) != ids.size()) {
+                         .where(ADMIN_AUTH.AUTH_ID.in(ids))
+                         .and(ADMIN_AUTH.ADMIN_ID.eq(admin))) != ids.size()) {
             throw new ServiceException(AuthorityConstant.NOT_EXIST);
         }
         return true;
